05-data-platforms/pipeline/assets/ingestion/requirements.txt
pandas==2.2.0
requests==2.31.0
pyarrow>=12.0.0
requests>=2.31.0
python-dateutil>=2.8.0
ingestr
tzdata
duckdb>=0.9.0